fix: move tree and proof methods to cbmt
move confusing methods to the cbmt type
