+++ examples/sha256/main.go
@@ -18,15 +18,14 @@ func main() {
 	for _, l := range leaves {
 		leavesI = append(leavesI, []byte(l))
 	}
-	mtree := merkle.Tree{
-		Nodes: leavesI,
+	cbmt := merkle.CBMT{
 		Merge: MergeByteArray{},
 	}
-	root := mtree.BuildRoot(leavesI)
+	root := cbmt.BuildMerkleRoot(leavesI)
 	fmt.Printf("Merkle root is %v \n", HashToStr(root))
 
 	// build merkle proof for 42 (its index is 1);
-	proof, err := mtree.BuildTreeAndProof(leavesI, []uint32{1})
+	proof, err := cbmt.BuildMerkleProof(leavesI, []uint32{1})
 	if err != nil {
 		panic(err)
 	}
@@ -45,7 +44,7 @@ func main() {
 	// fmt.Printf("merkle proof verify result is %v\n", verifyResult)
 
 	// build merkle proof for 42 and 20191116 (indices are 1 and 4);
-	proof, err = mtree.BuildTreeAndProof(leavesI, []uint32{1, 4})
+	proof, err = cbmt.BuildMerkleProof(leavesI, []uint32{1, 4})
 	if err != nil {
 		panic(err)
 	}
@@ -57,7 +56,7 @@ func main() {
 	fmt.Printf("merkle proof indices are %v\n", proof.Indices)
 
 	// retrieve leaves
-	retrievedLeaves, err := proof.RetriveLeaves(leavesI)
+	retrievedLeaves, err := cbmt.RetriveLeaves(proof, leavesI)
 	if err != nil {
 		panic(err)
 	}
