+++ merkle/proof.go
@@ -9,7 +9,7 @@ import (
 // GetRoot returns the root value of a merkle proof
 func (p *Proof) GetRoot(leaves []interface{}) (interface{}, error) {
 	if len(leaves) != len(p.Indices) || len(leaves) == 0 {
-		return 0, errors.New("leaves could not be empty")
+		return 0, errors.New("leaves count should equal to indices and not be empty")
 	}
 
 	// TODO: write function for interace sorting helpers.SortUint32Slice(leaves)
@@ -57,7 +57,7 @@ func (p *Proof) GetRoot(leaves []interface{}) (interface{}, error) {
 			Leaf:  parentNode,
 		})
 	}
-	return 0, errors.New("")
+	return 0, errors.New("root not found")
 }
 
 // Verify verifies the root value against tree leaves
@@ -70,35 +70,4 @@ func (p *Proof) Verify(root interface{}, leaves []interface{}) (bool, error) {
 		return true, nil
 	}
 	return false, nil
-}
-
-// RetriveLeaves returns the leaves of a merkle proof
-func (p Proof) RetriveLeaves(leaves []interface{}) ([]interface{}, error) {
-	if len(leaves) == 0 || len(p.Indices) == 0 {
-		return []interface{}{}, errors.New("leaves or indecies should not be empty")
-	}
-
-	leavesCount := uint32(len(leaves))
-	var validIndicesRange []uint32
-	for i := uint32(leavesCount - 1); i < uint32((leavesCount<<1)-1); i++ {
-		validIndicesRange = append(validIndicesRange, i)
-	}
-
-	var allProofsInRange = true
-
-	for _, v := range p.Indices {
-		if !helpers.Uint32SliceContains(validIndicesRange, v) {
-			allProofsInRange = false
-			break
-		}
-	}
-
-	var res []interface{}
-	if allProofsInRange {
-		for _, v := range p.Indices {
-			res = append(res, leaves[v+1-leavesCount])
-		}
-	}
-
-	return res, nil
 }
